Refactor Enhanced Table files
Cleaning up and renaming files in order to prepare support for Document
Table
